refactor: update import paths to use content directory for consistency
